pyproject.toml
[tool.poetry]
name = "luzia-web"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "luzia_web"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
streamlit = "^1.21.0"
pillow = "^9.5.0"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
